requirements.txt
Flask
Flask-Login
PyYAML
dnslib
sqlite3
email-validator
